Add darwin support and extend signals for windows
